Code cleanup  - Split Header component and remove search functionality  - perform general code cleanup
