+++ package.json
@@ -8,9 +8,11 @@
     "@testing-library/user-event": "^12.1.10",
     "bootstrap": "^4.6.0",
     "moment": "^2.29.1",
+    "moment-timezone": "^0.5.33",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.0",
     "react-dom": "^17.0.2",
+    "react-moment": "^1.1.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
